refactor: LoginModal을 features/auth로 이동
shared가 features/auth, features/profile을 참조하는 의존성 역전 수정.
features/profile 내부 경로 직접 참조를 index.ts 경유로 변경.
