+++ features/auth/index.ts
@@ -2,5 +2,7 @@ export { exchangeKakaoCode, fetchNeedsOnboarding } from './auth.api';
 export { LOGIN_RETURN_COOKIE_NAME } from './auth.constants';
 export { useLoginGate } from './login-modal.hooks';
 export { openLoginModalOnPage, resolveKakaoLoginDestination, saveLoginReturnPath } from './auth.lib';
+export { LandingHeader } from './components/landing-header';
 export { LoginButton } from './components/login-button';
+export { LoginModal } from './components/login-modal';
 export { SignupFlow } from './components/signup-flow';
